support arbitrary whitespace between elements
Previously format elements had to be separated by exactly one space
character. If there was a typo in the apache.conf that introduced more
spaces or a tab, this module couldn't parse the resulting format or log
file. Now the separator between log elements is matched by one or more
whitespace characters.

Also the return of a hashref from the parse method is now conditional on
a successful parse. A false value is returned if the parse fails.