Restrict multiple whitespace tolerance
The introduction of \s+ in between all elements led to a missing
non-quoted element being mistaken for a gap between two elements. Now
\s+ is only matched in between quoted elements where it is unambiguous.

Refactored tests so that lines and parsed hashes are adjacent, and
included tests for multiple spaces between quoted elements, and the
failure to parse due to a missing element.

Fixes RT#107511